+++ lib/android_key_attestation.rb
@@ -4,6 +4,9 @@ module AndroidKeyAttestation
   class Error < StandardError; end
   class ExtensionMissingError < Error; end
   class ChallengeMismatchError < Error; end
+  class CertificateVerificationError < Error; end
+
+  GEM_ROOT = File.expand_path(__dir__)
 end
 
 require_relative "android_key_attestation/statement"