+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `Statement#verify_certificate_chain` to verify if the attestation certificate is trustworthy
+
 ## [0.2.0] - 2019-12-31
 ### Changed
 - Raise `ChallengeMismatchError` if the challenge lengths are different, not `ArgumentError`